Added Floating-Point Instructions, Control Instructions, Channel Instructions
Hint-for-Branch Instructions are missing couldn't figure out how to parse brinst, brtarg. That means Hint for Branch (r-form), Hint for Branch (a-form), Hint for Branch Relative are not implemented. 
I've also checked InstructionsFactoryISDF and made some extra spaces in some OpCodes (mostly branch instructions with *D*CE****) in the SPU.isdf to make the reading easier.